requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr!=2.1.0,>=2.0.0 # Apache-2.0
oslo.config>=6.8.0 # Apache-2.0
openstacksdk>=0.48.0 # Apache-2.0
libvirt-python>=6.0.0 # LGPLv2+
pyghmi>=1.2.0 # Apache-2.0
cliff!=2.9.0,>=2.8.0 # Apache-2.0
pyzmq>=19.0.0  # LGPL+BSD
